CORS origins, landlords controller
In order to get the JavaScript debugger to cooperate, I added a new
origin in the cors initializer.

I placed some of components into a new Folder dir.
